work towards organizing outputs
