+++ fluviewer/fluviewer.py
@@ -16,7 +16,7 @@ import pandas as pd
 from . import __version__
 
 from . import cli_args
-from . import parsers
+from . import analysis
 from . import plots
 import fluviewer.logging
 
@@ -88,15 +88,27 @@ def main():
         
     log.info('Starting analysis...')
 
+    #
+    # Stage 0: Normalize depth of reads.
+    
     current_analysis_stage = 'normalize_depth'
     current_analysis_stage_index = analysis_stages.index(current_analysis_stage)
+    current_analysis_stage_outdir = os.path.join(args.outdir, 'analysis_by_stage', f'{current_analysis_stage_index:02}_{current_analysis_stage}')
+    current_analysis_stage_outdir = os.path.abspath(current_analysis_stage_outdir)
     log.info(f'Beginning analysis stage: {current_analysis_stage}')
+    log.info(f'Output directory: {current_analysis_stage_outdir}')
 
-    normalize_depth(
-        args.outdir,
+    current_analysis_stage_inputs = {
+        'input_reads_fwd': os.path.abspath(args.forward_reads),
+        'input_reads_rev': os.path.abspath(args.reverse_reads),
+    }
+
+    normalize_depth_analysis_summary = analysis.normalize_depth(
+        current_analysis_stage_inputs,
+        current_analysis_stage_outdir,
         args.output_name,
-        args.forward_reads,
-        args.reverse_reads,
+        os.path.abspath(args.forward_reads),
+        os.path.abspath(args.reverse_reads),
         args.target_depth,
         args.max_memory,
         args.disable_garbage_collection
@@ -105,40 +117,57 @@ def main():
     log.info(f'Analysis stage complete: {current_analysis_stage}')
 
 
+    #
+    # Stage 1: Assemble contigs.
+    
     current_analysis_stage = 'assemble_contigs'
     current_analysis_stage_index = analysis_stages.index(current_analysis_stage)
+    
+    current_analysis_stage_inputs = {
+        'normalized_reads_fwd': normalize_depth_analysis_summary['outputs']['normalized_reads_fwd'],
+        'normalized_reads_rev': normalize_depth_analysis_summary['outputs']['normalized_reads_rev'],
+    }
+    current_analysis_stage_outdir = os.path.join(args.outdir, 'analysis_by_stage', f'{current_analysis_stage_index:02}_{current_analysis_stage}')
+    current_analysis_stage_outdir = os.path.abspath(current_analysis_stage_outdir)
+
     log.info(f'Beginning analysis stage: {current_analysis_stage}')
 
-    assemble_contigs(
-        args.outdir,
+    assemble_contigs_analysis_summary = analysis.assemble_contigs(
+        current_analysis_stage_inputs,
+        current_analysis_stage_outdir,
         args.output_name,
         args.disable_garbage_collection
     )
 
     log.info(f'Analysis stage complete: {current_analysis_stage}')
 
+    #
+    # Stage 2: Blast contigs.
+
     current_analysis_stage = 'blast_contigs'
     current_analysis_stage_index = analysis_stages.index(current_analysis_stage)
+    current_analysis_stage_inputs = {
+        'contigs': assemble_contigs_analysis_summary['outputs']['contigs'],
+    }
+    current_analysis_stage_outdir = os.path.join(args.outdir, 'analysis_by_stage', f'{current_analysis_stage_index:02}_{current_analysis_stage}')
+    current_analysis_stage_outdir = os.path.abspath(current_analysis_stage_outdir)
     log.info(f'Beginning analysis stage: {current_analysis_stage}')
 
-    blast_results = blast_contigs(
-        args.database,
-        args.outdir,
+    blast_contigs_analysis_summary = analysis.blast_contigs(
+        current_analysis_stage_inputs,
+        current_analysis_stage_outdir,
         args.output_name,
         args.disable_garbage_collection,
+        args.database,
         args.threads,
         args.min_identity,
         args.min_alignment_length,
     )
 
-    filtered_blast_results = filter_contig_blast_results(
-        blast_results,
-        args.outdir,
-        args.output_name,
-        args.disable_garbage_collection
-    )
-
     log.info(f'Analysis stage complete: {current_analysis_stage}')
+
+    print(json.dumps(blast_contigs_analysis_summary, indent=4))
+    exit()
 
     current_analysis_stage = 'scaffolding'
     current_analysis_stage_index = analysis_stages.index(current_analysis_stage)
@@ -369,508 +398,11 @@ def check_database(db, outdir, out_name, collect_garbage):
             run(terminal_command, outdir, out_name, process_name, error_code, collect_garbage)
         except Exception as e:
             log.error(f'Error creating BLAST database: at {os.path.dirname(db)} {e}')
-            exit(1)
+            exit(1)    
 
 
-def run(terminal_command, outdir, out_name, process_name, error_code, collect_garbage):
-    """
-    A generalized function for running subprocesses, logging their output, and
-    trapping erroneous exit statuses.
-
-    :param terminal_command: The command to be run in the terminal.
-    :type terminal_command: str
-    :param outdir: Path to the output directory.
-    :type outdir: str
-    :param out_name: Name of the output directory.
-    :type out_name: str
-    :param process_name: Name of the process being run.
-    :type process_name: str
-    :param error_code: Exit status if the subprocess fails.
-    :type error_code: int
-    :param collect_garbage: Whether to remove intermediate files.
-    :type collect_garbage: bool
-    :return: None
-    :rtype: None
-    """
-    stdout_file = os.path.join(outdir, out_name, 'logs',
-                               f'{process_name}_stdout.txt')
-    stderr_file = os.path.join(outdir, out_name, 'logs',
-                               f'{process_name}_stderr.txt')
-    for file in [stdout_file, stderr_file]:
-        if file != None:
-            with open(file, 'w') as log_file:
-                log_file.write('*' * 80 + '\n')
-                log_file.write('Terminal command:' + '\n')
-                log_file.write(terminal_command + '\n')
-                log_file.write('*' * 80 + '\n')
-
-    complete_process = None
-    try:
-        with open(stdout_file, 'w') as stdout_file:
-            with open(stderr_file, 'w') as stderr_file:
-                complete_process = subprocess.run(
-                    terminal_command,
-                    stdout=stdout_file,
-                    stderr=stderr_file,
-                    shell=True
-                )
-    except Exception as e:
-        log.error(f'Error running subprocess {process_name}: {e}')
-        if collect_garbage:
-            garbage_collection(outdir, out_name)
-        exit(error_code)
-
-    return_code = complete_process.returncode
-    if return_code != 0:
-        log.error(f'Subprocess {process_name} failed (Exit status: {return_code})')
-        if collect_garbage:
-            garbage_collection(outdir, out_name)
-        exit(error_code)
 
 
-def garbage_collection(outdir, out_name):
-    """
-    Clean up unneccessary intermediate files (many of which occupy
-    substantial amounts of storage).
-
-    :param outdir: Path to the output directory.
-    :type outdir: str
-    :param out_name: Name of the output subdirectory.
-    :type out_name: str
-    :return: None
-    """
-    if os.path.isdir(os.path.join(outdir, out_name, f'spades_output')):
-        shutil.rmtree(os.path.join(outdir, out_name, f'spades_output'))
-
-    files = [
-        'R1.fq',
-        'R2.fq',
-        'contigs_blast.tsv',
-        'scaffolds.fa',
-        'scaffolds_blast.tsv',
-        'alignment.sam',
-        'pileup.vcf',
-        'variants.bcf',
-        'variants.bcf.csi',
-        'low_cov.tsv',
-        'ambig.tsv',
-        'variants.tsv',
-        'masked.bed',
-        'reads_mapped.tsv',
-        'depth_of_cov_samtools.tsv',
-        'depth_of_cov_freebayes.tsv',
-    ]
-
-    bwa_suffixes = ['amb', 'ann', 'bwt', 'fai', 'pac', 'sa']
-    for suffix in bwa_suffixes:
-        files.append(f'{out_name}_mapping_refs.fa.{suffix}')
-
-    segments = 'PB2 PB1 PA HA NP NA M NS'.split(' ')
-    for segment in segments:
-        files += [
-            f'{segment}_contigs.fa',
-            f'{segment}_contigs.afa',
-            f'{segment}_contigs.dnd'
-        ]
-
-    for file in files:
-        file = os.path.join(out_name, file)
-        if os.path.isfile(file):
-            os.remove(file)
-            log.info(f'Removed intermediate file: {file}')
-
-
-def normalize_depth(
-        outdir: Path,
-        out_name: str,
-        fwd_reads_raw: Path,
-        rev_reads_raw: Path,
-        depth: int,
-        max_memory: int,
-        collect_garbage: bool
-):
-    """
-    BBNorm is run on the input reads to downsample regions of deep coverage
-    (using a k-mer frequency approach). This balances coverage, increases
-    analysis speed, and limits the impacts of artefactual reads.
-
-    :param outdir: Path to the output directory.
-    :type outdir: str
-    :param out_name: Name of the output directory.
-    :type out_name: str
-    :param fwd_reads_raw: Path to the raw forward reads.
-    :type fwd_reads_raw: str
-    :param rev_reads_raw: Path to the raw reverse reads.
-    :type rev_reads_raw: str
-    :param depth: Target depth of coverage.
-    :type depth: int
-    :param max_memory: Maximum memory to allocate to BBNorm.
-    :type max_memory: int
-    :param collect_garbage: Whether to remove intermediate files.
-    :type collect_garbage: bool
-    """
-    log.info('Normalizing depth of coverage and subsampling reads...')
-    fwd_reads = os.path.join(outdir, out_name, f'R1.fq')
-    rev_reads = os.path.join(outdir, out_name, f'R2.fq')
-    terminal_command = (f'bbnorm.sh in={fwd_reads_raw} in2={rev_reads_raw} '
-                        f'out={fwd_reads} out2={rev_reads} target={depth}')
-    terminal_command = (terminal_command + f' -Xmx{max_memory}g'
-                        if max_memory is not None else terminal_command)
-    process_name = 'bbnorm'
-    error_code = 2
-    run(terminal_command, outdir, out_name, process_name, error_code, collect_garbage)
-
-    bbnorm_stderr_log_path = os.path.join(outdir, out_name, 'logs', 'bbnorm_stderr.txt')
-    bbnorm_log = parsers.parse_bbnorm_log(bbnorm_stderr_log_path)
-
-    for pass_num, pass_stats in bbnorm_log.items():
-        if not pass_num.startswith('pass_'):
-            continue
-        pass_num_int = int(pass_num.split('_')[-1])
-        log.info(f'Normalization pass {pass_num_int}: Total reads in: {pass_stats["total_reads_in"]}')
-        log.info(f'Normalization pass {pass_num_int}: Percent reads kept: {pass_stats["total_reads_kept_percent"]}%')
-        log.info(f'Normalization pass {pass_num_int}: Percent unique: {pass_stats["percent_unique"]}%')
-        log.info(f'Normalization pass {pass_num_int}: Average depth (unique kmers): {pass_stats["depth_average_unique_kmers"]}')
-        log.info(f'Normalization pass {pass_num_int}: Average depth (all kmers): {pass_stats["depth_average_all_kmers"]}')
-        log.info(f'Normalization pass {pass_num_int}: Approx. median read depth: {pass_stats["approx_read_depth_median"]}')
-
-    with open(os.path.join(outdir, out_name, 'logs', 'bbnorm_log.json'), 'w') as f:
-        json.dump(bbnorm_log, f, indent=4)
-        f.write('\n')
-    
-
-    
-def assemble_contigs(outdir: Path, out_name: str, collect_garbage: bool):
-    """
-    Normalized, downsampled reads are assembled de novo into contigs
-    using SPAdes.
-
-    :param outdir: Path to the output directory.
-    :type outdir: Path
-    :param out_name: Name used for outputs
-    :type out_name: str
-    :param collect_garbage: Whether to remove intermediate files.
-    :type collect_garbage: bool
-    """
-    log.info('Assembling reads into contigs...')
-    spades_output = os.path.join(outdir, out_name, 'spades_output')
-    fwd_reads = os.path.join(outdir, out_name, f'R1.fq')
-    rev_reads = os.path.join(outdir, out_name, f'R2.fq')
-    terminal_command = (f'spades.py --rnaviral --isolate -1 {fwd_reads} '
-                        f'-2 {rev_reads} -o {spades_output}')
-    process_name = 'spades'
-    error_code = 3
-    run(terminal_command, outdir, out_name, process_name, error_code, collect_garbage)
-    if not os.path.isfile(os.path.join(spades_output, 'contigs.fasta')):
-        log.error('No contigs assembled! Aborting analysis.')
-        if collect_garbage:
-            garbage_collection(out_name)
-        error_code = 4
-        exit(error_code)
-    else:
-        num_contigs = 0
-        with open(os.path.join(spades_output, 'contigs.fasta'), 'r') as f:
-            for line in f:
-                if line.startswith('>'):
-                    num_contigs += 1
-        log.info('Contigs assembled successfully.')
-        log.info(f'Assembled {num_contigs} contigs.')
-
-
-def blast_contigs(db, outdir, out_name, collect_garbage, threads, identity, length):
-    """
-    Contigs are aligned to reference sequences using BLASTn.
-
-    :param db: Path to the reference sequence database.
-    :type db: Path
-    :param out_name: Name used for outputs.
-    :type out_name: str
-    :param collect_garbage: Whether to remove intermediate files.
-    :type collect_garbage: bool
-    :param threads: Number of threads to use for BLASTn.
-    :type threads: int
-    :param identity: Minimum sequence identity for BLASTn hits.
-    :type identity: float
-    :param length: Minimum alignment length for BLASTn hits.
-    :type length: int
-    :return: BLASTn results.
-    :rtype: pd.DataFrame
-    """
-    log.info('Aligning contigs to reference sequences...')
-    db = os.path.abspath(db)
-    blast_output = os.path.join(outdir, out_name, 'contigs_blast.tsv')
-    spades_output = os.path.join(outdir, out_name, 'spades_output')
-    contigs = os.path.join(spades_output, 'contigs.fasta')
-    cols = [
-        'qseqid',
-        'sseqid',
-        'pident',
-        'length',
-        'bitscore',
-        'sstart',
-        'send',
-        'qseq',
-        'sseq',
-        'slen',
-    ]
-    cols_str = ' '.join(cols)
-    terminal_command = (f'blastn -query {contigs} -db {db} '
-                        f'-num_threads {threads} -outfmt "6 {cols_str}" '
-                        f'> {blast_output}')
-    process_name = 'blastn_contigs'
-    error_code = 6
-    run(terminal_command, outdir, out_name, process_name, error_code, collect_garbage)
-    blast_results = pd.read_csv(blast_output, names=cols, sep='\t')
-    total_num_blast_results = len(blast_results)
-    log.info('Contigs aligned to reference sequences.')
-    log.info(f'Found {total_num_blast_results} total matches.')
-    blast_results = blast_results[blast_results['pident']>=identity]
-    blast_results = blast_results[blast_results['length']>=length]
-    num_blast_results_after_identity_and_length_filter = len(blast_results)
-    log.info(f'Found {num_blast_results_after_identity_and_length_filter} matches with at least {identity}% identity and {length} bp alignment length.')
-    percent_blast_results_retained = num_blast_results_after_identity_and_length_filter / total_num_blast_results * 100
-    log.info(f'Retained {percent_blast_results_retained:.2f}% matches for further analysis.')
-
-    if len(blast_results) == 0:
-        log.error(f'No contigs aligned to reference sequences! Aborting analysis.')
-        if collect_garbage:
-            garbage_collection(outdir, out_name)
-        error_code = 7
-        exit(error_code)
-
-    return blast_results
-
-
-def filter_contig_blast_results(blast_results, outdir, out_name, collect_garbage):
-    """
-    Contigs alignments are filtered to discard spurious alignments.
-
-    The length and sequence identity of each alignment must exceed certain
-    thresholds.
-
-    Afterwards, a single reference sequence is selected for each
-    genome segment. The reference sequence with the most positions covered by
-    contigs. Ties are broken by:
-   
-      1. The highest sequence identity
-      2. The longest reference sequence length
-      3. First alphabetically
-
-    Once a reference sequence has been chosen for each segment, only contig alignments
-    to those reference sequences are retained.
-
-    :param blast_results: BLASTn results.
-    :type blast_results: pd.DataFrame
-    :param out_name: Name used for outputs.
-    :type out_name: str
-    :param collect_garbage: Whether to remove intermediate files.
-    :type collect_garbage: bool
-    :return: Filtered BLASTn results.
-    :rtype: pd.DataFrame
-    """
-    log.info('Filtering contig alignments...')
-    # 
-    # Annotate each ref seq with its segment and subtype.
-    subject_annots = blast_results[['sseqid']].drop_duplicates()
-
-    # sseqid format: accession|strain_name|segment|subtype
-    get_segment = lambda row: row['sseqid'].split('|')[2]
-    subject_annots['segment'] = subject_annots.apply(get_segment, axis=1)
-
-    get_subtype = lambda row: row['sseqid'].split('|')[3]
-    subject_annots['subtype'] = subject_annots.apply(get_subtype, axis=1)
-
-    # Merge segment and subtype annotations back into blast_results
-    blast_results = pd.merge(blast_results, subject_annots, on='sseqid')
-
-    # Check for evidence of mixed infections. First, find best alignments(s)
-    # for each contig. Next, check if each segment is represented by only one subtype.
-    cols = ['qseqid', 'bitscore']
-    # Find best alignment(s) for each contig, based on bitscore
-    max_bitscores = blast_results[cols].groupby('qseqid').max().reset_index()
-    # Use the qseqid and bitscore to find the best alignment(s) for each contig
-    best_results = pd.merge(blast_results, max_bitscores, on=cols)
-
-    log.info('Checking for mixed infections...')
-    # Check for multiple subtypes for each segment
-    segments = blast_results['segment'].unique()
-    for segment in segments:
-        log.info(f'Checking segment {segment}...')
-        segment_results = best_results[best_results['segment']==segment]
-        log.info(f'Found {len(segment_results)} contigs for segment {segment}.')
-        segment_subtypes = segment_results['subtype'].unique()
-        if len(segment_subtypes) > 1:
-            segment_subtypes = ', '.join(segment_subtypes)
-            log.error(f'Multiple subtypes detected for segment {segment} '
-                      f'({segment_subtypes})! Aborting analysis.\n')
-            if collect_garbage:
-                garbage_collection(out_name)
-            error_code = 8
-            exit(error_code)
-        else:
-            if segment_subtypes[0] == 'none':
-                log.info(f'No subtype determined for segment {segment}.')
-            else:
-                log.info(f'Subtype determined for segment {segment}: {segment_subtypes[0]}.')
-
-    #
-    # Find ref seq(s) most covered by contigs.
-    log.info('Finding reference sequences most covered by contigs...')
-    def count_cov_pos(data_frame):
-        """
-        Count the number of positions covered by contigs.
-        Iterates through each alignment, and adds the range of positions
-        covered by the alignment to a set. The length of the set is the number
-        of covered positions.
-
-        :param data_frame: DataFrame containing BLASTn results.
-        :type data_frame: pd.DataFrame
-        :return: Number of covered positions.
-        :rtype: int
-        """
-        cov_positions = set()
-        for index, row in data_frame.iterrows():
-            start = min([row['sstart'], row['send']])
-            end = max([row['sstart'], row['send']])
-            cov_positions = cov_positions.union(set(range(start, end + 1)))
-        return len(cov_positions)
-
-    cols = [
-        'sseqid',
-        'segment',
-        'subtype',
-        'sstart',
-        'send'
-    ]
-    group_cols = [
-        'sseqid',
-        'segment',
-        'subtype'
-    ]
-    cov_pos = blast_results[cols].drop_duplicates()
-
-    # Count the number of positions covered by contigs for each ref seq
-    cov_pos = cov_pos.groupby(group_cols).apply(count_cov_pos).reset_index()
-    log.info('Found covered positions for each reference sequence.')
-    
-    cov_pos.columns = [
-        'sseqid',
-        'segment',
-        'subtype',
-        'covered_positions'
-    ]
-    cols = [
-        'segment',
-        'subtype',
-        'covered_positions'
-    ]
-    group_cols = ['segment', 'subtype']
-
-    # Find the ref seq(s) with the most covered positions for each segment/subtype
-    max_cov_pos = cov_pos[cols].drop_duplicates()
-    max_cov_pos = max_cov_pos.groupby(group_cols).max().reset_index()
-    log.info('Found reference sequence(s) with the most covered positions for each segment/subtype.')
-    merge_cols = ['segment', 'subtype', 'covered_positions']
-    max_cov_pos = pd.merge(cov_pos, max_cov_pos, on=merge_cols)
-
-    # Log a summary of covered positions for each segment/subtype
-    for segment in segments:
-        segment_results = max_cov_pos[max_cov_pos['segment']==segment]
-        subtype_to_log = '(undetermined)'
-        if segment_results['subtype'].values[0] != 'none':
-            subtype_to_log = segment_results['subtype'].values[0]
-        for subtype in segment_results['subtype'].unique():
-            num_covered_positions = segment_results[segment_results['subtype']==subtype]['covered_positions'].values[0]
-            log.info(f'Segment: {segment}, Subtype: {subtype_to_log}: {num_covered_positions} covered positions.')
-
-    max_cov_pos = max_cov_pos[['sseqid', 'covered_positions']]
-    log.info('Found reference sequence(s) with the most covered positions for each segment/subtype.')
-    blast_results = pd.merge(blast_results, max_cov_pos, on='sseqid')
-    num_blast_results = len(blast_results)
-    log.info(f'Found {num_blast_results} total contig alignments.')
-
-    #
-    # Find remaining ref seq(s) with most identical positions.
-    def count_id_pos(data_frame):
-        """
-        Count the number of identical positions between contigs and reference.
-        Determined by comparing the query and subject sequences from BLASTn.
-        Iterates through each base in the query sequence, and if the base is
-        a nucleotide and matches the subject sequence, the subject position is
-        added to a set. The length of the set is the number of identical
-        positions.
-
-        :param data_frame: DataFrame containing BLASTn results.
-        :type data_frame: pd.DataFrame
-        :return: Number of identical positions.
-        :rtype: int
-        """
-        identical_positions = set()
-        for index, row in data_frame.iterrows():
-            start = min([row['sstart'], row['send']])
-            increment = 1 if row['sstart'] <= row['send'] else -1
-            subject_position = start
-            for qbase, sbase in zip(row['qseq'], row['sseq']):
-                if sbase in 'ATGC' and qbase == sbase:
-                    identical_positions.add(subject_position)
-                if sbase != '-':
-                    subject_position += increment
-        return len(identical_positions)
-
-    cols = [
-        'sseqid',
-        'segment',
-        'subtype',
-        'sstart',
-        'send',
-        'qseq',
-        'sseq'
-    ]
-    group_cols = ['sseqid', 'segment', 'subtype']
-    ident_pos = blast_results[cols].drop_duplicates()
-    # Count the number of identical positions for each ref seq
-    ident_pos = ident_pos.groupby(group_cols).apply(count_id_pos).reset_index()
-    ident_pos.columns = ['sseqid', 'segment', 'subtype', 'identical_positions']
-    cols = ['segment', 'subtype', 'identical_positions']
-    group_cols = ['segment', 'subtype']
-
-    # Find the ref seq(s) with the most identical positions for each segment/subtype
-    max_ident_pos = ident_pos[cols].drop_duplicates()
-    max_ident_pos = max_ident_pos.groupby(group_cols).max().reset_index()
-    merge_cols = ['segment', 'subtype', 'identical_positions']
-    max_ident_pos = pd.merge(ident_pos, max_ident_pos, on=merge_cols)
-    log.info('Found reference sequence(s) with the most identical positions for each segment/subtype.')
-    for segment in segments:
-        segment_results = max_ident_pos[max_ident_pos['segment']==segment]
-        for subtype in segment_results['subtype'].unique():
-            subtype_to_log = subtype if subtype != 'none' else 'undetermined'
-            num_identical_positions = segment_results[segment_results['subtype']==subtype]['identical_positions'].values[0]
-            log.info(f'Segment {segment}, Subtype {subtype_to_log}: {num_identical_positions} identical positions.')
-
-    cols = ['sseqid', 'identical_positions']
-    max_ident_pos = max_ident_pos[cols].drop_duplicates()
-    blast_results = pd.merge(blast_results, max_ident_pos, on='sseqid')
-    
-    # Take longest remaining ref seq for each segment/subtype.
-    cols = ['segment', 'subtype', 'slen']
-    group_cols = ['segment', 'subtype']
-    longest_sseq = blast_results[cols].drop_duplicates()
-    longest_sseq = longest_sseq.groupby(group_cols).min().reset_index()
-    blast_results = pd.merge(blast_results, longest_sseq, on=cols)
-
-    # Take first alphabetical remaining ref seq for each segment/subtype.
-    cols = ['segment', 'subtype', 'sseqid']
-    group_cols = ['segment', 'subtype']
-    first_alpha_sseq = blast_results[cols].drop_duplicates()
-    first_alpha_sseq = first_alpha_sseq.groupby(group_cols).min().reset_index()
-    blast_results = pd.merge(blast_results, first_alpha_sseq, on=cols)
-    for segment in segments:
-        segment_results = blast_results[blast_results['segment']==segment]
-        for subtype in segment_results['subtype'].unique():
-            subtype_to_log = subtype if subtype != 'none' else '(undetermined)'
-            ref_seq = segment_results[segment_results['subtype']==subtype]['sseqid'].values[0]
-            log.info(f'Segment {segment}, Subtype: {subtype_to_log}. Selected reference sequence: {ref_seq}')
-    return blast_results
 
 
 def make_scaffold_seqs(blast_results, outdir, out_name, collect_garbage):
